+++ package.json
@@ -21,7 +21,8 @@
     "@walletconnect/utils": "2.19.1",
     "ethers": "^6.13.5",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.9.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.22.0",
