+++ src/App.css
@@ -222,3 +222,14 @@ a {
 hr {
   margin: 2rem 0;
 }
+
+.config {
+  margintop: '10px';
+  padding: '5px 10px';
+  fontsize: '12px';
+  backgroundcolor: '#666';
+  color: 'white';
+  border: 'none';
+  borderradius: '4px';
+  cursor: 'pointer';
+}
